Remove --disable-banner option from main command again
